# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend TixIT",
        "title": "TixIT",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/movies": {
            "get": {
                "description": "untuk mendapatkan list semua movie",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "get all list movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.MovieData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response401"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "untuk membuat movie baru",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "create movie",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": " ",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": " ",
                        "name": "banner",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "tag",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "releaseDate",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "duration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": " ",
                        "name": "synopsis",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Movie"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response401"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "untuk mendapatkan list movie dari id",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "get list movie by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "movie id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Movie"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response401"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "untuk menghapus movie",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "delete movie by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "movie id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Movie"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response401"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "pageInfo": {},
                "results": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "controllers.Response401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "default": "Unauthorized"
                },
                "success": {
                    "type": "boolean",
                    "default": false
                }
            }
        },
        "models.Movie": {
            "type": "object",
            "properties": {
                "castName": {
                    "type": "string"
                },
                "directorName": {
                    "type": "string"
                },
                "duration": {
                    "description": "ReleaseDate time.Time `json:\"releaseDate\" form:\"release_date\"`",
                    "type": "string"
                },
                "genreName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "synopsis": {
                    "type": "string"
                },
                "tag": {
                    "description": "Image *multipart.FileHeader `json:\"image\" form:\"image\" binding:\"required\"`\nBanner *multipart.FileHeader `json:\"banner\" form:\"banner\" binding:\"required\"`",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.MovieData": {
            "type": "object",
            "required": [
                "banner",
                "image"
            ],
            "properties": {
                "banner": {
                    "type": "string"
                },
                "castName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "directorName": {
                    "type": "string"
                },
                "duration": {
                    "description": "ReleaseDate time.Time `json:\"releaseDate\" form:\"release_date\"`",
                    "type": "string"
                },
                "genreName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string"
                },
                "tag": {
                    "description": "Image *multipart.FileHeader `json:\"image\" form:\"image\" binding:\"required\"`\nBanner *multipart.FileHeader `json:\"banner\" form:\"banner\" binding:\"required\"`",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}